feat(react) вынес useInitLaunchAnalyticsValues в отдельный хук для инициализации + добавил useLaunchAnalytics для инициализации работы launch события
