+++ packages/react/src/index.ts
@@ -3,4 +3,6 @@ export * from './useAddPlainData';
 export * from './useBlockRef';
 export * from './useElementRef';
 export * from './useElementShowRef';
+export * from './useInitLaunchAnalyticsValues';
+export * from './useLaunchAnalytics';
 export * from './usePageAnalytics';
